Portfolio holding summaries and trades (#71)
* Show summary of portfolio holdings

* Show a holding's individual share trades

* Allow fractional quantities in share trades

+++ share-module/share-service/src/main/java/com/hillayes/shares/domain/ShareTradeSummary.java
@@ -25,7 +25,7 @@ public class ShareTradeSummary {
     @ToString.Include
     private String name;
 
-    private long quantity;
+    private BigDecimal quantity;
 
     private BigDecimal totalCost;
